requirements.txt
streamlit
openai
google-cloud-aiplatform
google-auth
